Changes to build with Mono/MonoDevelop on Linux. Requires Mono 2.10 and MonoDevelop 2.6 beta 3+

+++ GameServer/gameobjects/GamePlayer.cs
@@ -6732,7 +6732,7 @@ namespace DOL.GS
 
 							if (spell != null)
 							{
-								int chance = reactiveItem.ProcChance > 0 ? reactiveItem.ProcChance : 10;
+								int chance = reactiveItem.ProcChance > 0 ? (int)reactiveItem.ProcChance : 10;
 
 								if (Util.Chance(chance))
 								{
@@ -6751,7 +6751,7 @@ namespace DOL.GS
 
 							if (spell != null)
 							{
-								int chance = reactiveItem.ProcChance > 0 ? reactiveItem.ProcChance : 10;
+								int chance = reactiveItem.ProcChance > 0 ? (int)reactiveItem.ProcChance : 10;
 
 								if (Util.Chance(chance))
 								{
